requirements.txt
annotated-types==0.6.0
astroid==3.0.1
autopep8==2.0.4
blinker==1.7.0
Cerberus==1.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.7
distlib==0.3.7
docopt==0.6.2
dropbox==11.36.2
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
gunicorn==21.2.0
idna==3.6
isort==5.13.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
pep517==0.13.1
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==4.1.0
plette==0.4.4
ply==3.11
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pylint==3.0.2
python-dotenv==1.0.0
requests==2.31.0
requirementslib==3.0.0
six==1.16.0
SQLAlchemy==2.0.23
stone==3.3.1
tomlkit==0.12.3
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
yarg==0.1.9
